create new FunctionManager, add push() and pop() methods to ss
The addition function now uses the new mechanism. The next steps are:

1) Convert all other operations to this system.
2) Consider whether maybe even stack operations should use it. Obviously
number entry, backspace, &c will not, however.
3) Delete the legacy code.

We are changing from manually overwriting space on the stack (which was
maybe partially optimized but unnecessarily complicated) to calling the
redraw method after every operation. As I understand it, curses will
work out which things actually need to be *changed* by that, so we won't
have the significantly greater (and maybe actually noticeable) latency
of updating things on the terminal that don't need to be updated, just
the overhead of having the program *tell* curses to update them.