finish moving all functions over to new system
This did break some things that worked before:
- Undo never works on anything anymore, whereas it sometimes did
something useful before.
- The status bar never displays errors now.
- The commands window no longer changes to display the current commands.
- There is no error checking for anything.

These are, of course, the next priorities. We may want to take a peek at
the deleted, old code to see how to manage each of these things.

The new abstraction is wonderful, however; with the proper
FunctionManager code in place, it is ridiculous how easy the previously
difficult operations were. Especially seeing that my long special-case
stack operations could be replaced with a single lambda function...