menu titles; division by zero
In the process of fixing the latter, I noticed that ss.push() had been
written wrong: it pushed multiple arguments in the wrong order. This
didn't get noticed because the only place it was used was for 'x', and I
had determined which item was which argument by trial and error.