beginner's programming documentation
Not sure how many people to whom this would apply will actually use esc
(probably zero to few), but you never know, and it was a good exercise
to write, one which also led to me writing down a few bugs and fixing a
few particularly bad ones.

I still need to write the more technical version of this, which will be
called (IIRC) `devel.py`.