refactor main function
Split out into a number of useful functions. In order to accomplish
this, we also changed the HistoricalStack to use a Memento pattern so
that an existing object can be restored rather than replaced with a
different object entirely (this just made it possible to avoid multiple
ugly returns from a function and does a neater job encapsulating the
changes).