invert order of pushing and popping
This is much more intuitive and I don't know why I ever did it any other
way. Before, 0 was the "bottom" of the stack, which was weird. This way,
if you request 3 items, you get the 3 items in the order they were
placed on the stack, with third on stack as [0] and bos as [2]. So
saying lambda i: i[0] - i[1] means subtract bos from sos, as you'd
normally read.