add function to sum entire stack
I often find myself spamming '+' when using esc, so this should be a
useful function. It's also the impetus for a nice option for log_as
where instead of a format string it is a callable which takes args and
retval and returns a string to log. That's needed because if you don't
know up-front how many args you're getting, it's impossible to produce a
correct positional format string ({-1} isn't a valid placeholder).