truncate history items instead of overwrapping
We will want some way to see the entirety of them, or they're not super
useful (still better than nothing, since they give you some anchoring
and you can always undo to them to see the details if really necessary).
We'll think about the best way to handle that -- maybe if there's a mode
for scrolling through the history, there could be a pop-up or something.