create basic help display based on actual function info
Includes the signature and description info (the docstring of the
function). I would like to see it also do a simulation of what would
happen if you did the calculation right now (we might need an option to
turn this off if your custom function is going to take a really long
time, or worse have side effects -- which it sure shouldn't, but better
safe than sorry).

We also need to make it so you can actually choose what you want to see
help on -- right now it is hard-coded. And the help display doesn't work
for menus. I think we need to make a Help object of some kind or in some
way seriously rethink this architecture, because the number of
properties and parameters is exploding out of control.