don't hide the stack when showing function help
The stack is much more useful than the commands window when viewing the
help of a menu.

For menus, though, the commands stay visible instead of the stack, as
those are more useful. We need to make it allow recursing into there now
so we can actually view the help of items on menus.