allow getting help on builtins
The stub classes offer some interesting possibilities: for instance, we
might want to move some of the actual processing of builtin functions
over there. The struggle would be breaking the interface of the
EscFunction class in a somewhat ugly way, since they would each need
different values -- so we might need a series of different execute
functions, for instance.