fix all Sphinx warnings and nitpicks
Mostly these were broken references. It took me about an hour to figure
out that even if a reference is to an existing Python object, it (of
course) will be a broken reference if that object is not *documented*. I
say "of course" because in retrospect obviously there's nowhere for the
link to go in that case, but it's actually not all that intuitive and
definitely bears remembering so I don't waste that much time again in
the future.

Since we are not going for comprehensive documentation of every class, I
simply went away from using, e.g., :class:`foo` and started doing
``foo`` when I never expect to have the documentation for foo in Sphinx.
This seems a little ugly, but it ought to be a nice compromise since we
wouldn't be able to use nitpicky mode otherwise and having it on allows
us to spot places we actually expected the reference target to exist,
which we would likely never notice otherwise.